correction erreur taille adresse

+++ src/process.h
@@ -8,8 +8,8 @@
  *  sp : "stack pointer", pointeur de pile
  */
 struct ctx_s {
-    uint16_t pc;
-    uint16_t sp;
+    uint32_t pc;
+    uint32_t sp;
 };
 
 /* Initialise un contexte */
